refactor: More restructure
